strutil,osutil: move kernel commandline helpers to osutil
